+++ lib/auth/permissions_test.go
@@ -22,51 +22,108 @@ import (
 	"time"
 
 	"github.com/gravitational/trace"
-	"github.com/jonboulle/clockwork"
 	"github.com/stretchr/testify/require"
 
 	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/lib/backend/memory"
+	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/services/local"
 	"github.com/gravitational/teleport/lib/tlsca"
 )
 
+const clusterName = "test-cluster"
+
 func TestContextLockTargets(t *testing.T) {
 	t.Parallel()
-	authContext := &Context{
-		Identity: BuiltinRole{
-			Role:        types.RoleNode,
-			ClusterName: "cluster",
-			Identity: tlsca.Identity{
-				Username: "node.cluster",
-				Groups:   []string{"role1", "role2"},
+
+	tests := []struct {
+		role types.SystemRole
+		want []types.LockTarget
+	}{
+		{
+			role: types.RoleNode,
+			want: []types.LockTarget{
+				{Node: "node", ServerID: "node"},
+				{Node: "node.cluster", ServerID: "node.cluster"},
+				{User: "node.cluster"},
+				{Role: "role1"},
+				{Role: "role2"},
+				{Role: "mapped-role"},
 			},
 		},
-		UnmappedIdentity: WrapIdentity(tlsca.Identity{
-			Username: "node.cluster",
-			Groups:   []string{"mapped-role"},
-		}),
+		{
+			role: types.RoleAuth,
+			want: []types.LockTarget{
+				{ServerID: "node"},
+				{ServerID: "node.cluster"},
+				{User: "node.cluster"},
+				{Role: "role1"},
+				{Role: "role2"},
+				{Role: "mapped-role"},
+			},
+		},
+		{
+			role: types.RoleProxy,
+			want: []types.LockTarget{
+				{ServerID: "node"},
+				{ServerID: "node.cluster"},
+				{User: "node.cluster"},
+				{Role: "role1"},
+				{Role: "role2"},
+				{Role: "mapped-role"},
+			},
+		},
+		{
+			role: types.RoleKube,
+			want: []types.LockTarget{
+				{ServerID: "node"},
+				{ServerID: "node.cluster"},
+				{User: "node.cluster"},
+				{Role: "role1"},
+				{Role: "role2"},
+				{Role: "mapped-role"},
+			},
+		},
+		{
+			role: types.RoleDatabase,
+			want: []types.LockTarget{
+				{ServerID: "node"},
+				{ServerID: "node.cluster"},
+				{User: "node.cluster"},
+				{Role: "role1"},
+				{Role: "role2"},
+				{Role: "mapped-role"},
+			},
+		},
 	}
-	expected := []types.LockTarget{
-		{Node: "node"},
-		{Node: "node.cluster"},
-		{User: "node.cluster"},
-		{Role: "role1"},
-		{Role: "role2"},
-		{Role: "mapped-role"},
+	for _, tt := range tests {
+		t.Run(tt.role.String(), func(t *testing.T) {
+			authContext := &Context{
+				Identity: BuiltinRole{
+					Role:        tt.role,
+					ClusterName: "cluster",
+					Identity: tlsca.Identity{
+						Username: "node.cluster",
+						Groups:   []string{"role1", "role2"},
+					},
+				},
+				UnmappedIdentity: WrapIdentity(tlsca.Identity{
+					Username: "node.cluster",
+					Groups:   []string{"mapped-role"},
+				}),
+			}
+			require.ElementsMatch(t, authContext.LockTargets(), tt.want)
+		})
 	}
-	require.ElementsMatch(t, authContext.LockTargets(), expected)
 }
 
 func TestAuthorizeWithLocksForLocalUser(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
 
-	srv, err := NewTestAuthServer(TestAuthServerConfig{
-		Dir:   t.TempDir(),
-		Clock: clockwork.NewFakeClock(),
-	})
-	require.NoError(t, err)
+	client, watcher, authorizer := newTestResources(t)
 
-	user, role, err := CreateUserAndRole(srv.AuthServer, "test-user", []string{})
+	user, role, err := createUserAndRole(client, "test-user", []string{}, nil)
 	require.NoError(t, err)
 	localUser := LocalUser{
 		Username: user.GetName(),
@@ -83,15 +140,15 @@ func TestAuthorizeWithLocksForLocalUser(t *testing.T) {
 		Target: types.LockTarget{MFADevice: localUser.Identity.MFAVerified},
 	})
 	require.NoError(t, err)
-	upsertLockWithPutEvent(ctx, t, srv, mfaLock)
+	upsertLockWithPutEvent(ctx, t, client, watcher, mfaLock)
 
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
+	_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
 	require.Error(t, err)
 	require.True(t, trace.IsAccessDenied(err))
 
 	// Remove the MFA record from the user value being authorized.
 	localUser.Identity.MFAVerified = ""
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
+	_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
 	require.NoError(t, err)
 
 	// Add an access request lock.
@@ -99,16 +156,16 @@ func TestAuthorizeWithLocksForLocalUser(t *testing.T) {
 		Target: types.LockTarget{AccessRequest: localUser.Identity.ActiveRequests[0]},
 	})
 	require.NoError(t, err)
-	upsertLockWithPutEvent(ctx, t, srv, requestLock)
+	upsertLockWithPutEvent(ctx, t, client, watcher, requestLock)
 
 	// localUser's identity with a locked access request is locked out.
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
+	_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
 	require.Error(t, err)
 	require.True(t, trace.IsAccessDenied(err))
 
 	// Not locked out without the request.
 	localUser.Identity.ActiveRequests = nil
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
+	_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
 	require.NoError(t, err)
 
 	// Create a lock targeting the role written in the user's identity.
@@ -116,9 +173,9 @@ func TestAuthorizeWithLocksForLocalUser(t *testing.T) {
 		Target: types.LockTarget{Role: localUser.Identity.Groups[0]},
 	})
 	require.NoError(t, err)
-	upsertLockWithPutEvent(ctx, t, srv, roleLock)
+	upsertLockWithPutEvent(ctx, t, client, watcher, roleLock)
 
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
+	_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, localUser))
 	require.Error(t, err)
 	require.True(t, trace.IsAccessDenied(err))
 }
@@ -127,42 +184,41 @@ func TestAuthorizeWithLocksForBuiltinRole(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
 
-	srv, err := NewTestAuthServer(TestAuthServerConfig{
-		Dir:   t.TempDir(),
-		Clock: clockwork.NewFakeClock(),
-	})
-	require.NoError(t, err)
+	client, watcher, authorizer := newTestResources(t)
+	for _, role := range types.LocalServiceMappings() {
+		t.Run(role.String(), func(t *testing.T) {
+			builtinRole := BuiltinRole{
+				Username: "node",
+				Role:     role,
+				Identity: tlsca.Identity{
+					Username: "node",
+				},
+			}
 
-	builtinRole := BuiltinRole{
-		Username: "node",
-		Role:     types.RoleNode,
-		Identity: tlsca.Identity{
-			Username: "node",
-		},
+			// Apply a node lock.
+			nodeLock, err := types.NewLock("node-lock", types.LockSpecV2{
+				Target: types.LockTarget{ServerID: builtinRole.Identity.Username},
+			})
+			require.NoError(t, err)
+			upsertLockWithPutEvent(ctx, t, client, watcher, nodeLock)
+
+			_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, builtinRole))
+			require.Error(t, err)
+			require.True(t, trace.IsAccessDenied(err))
+
+			builtinRole.Identity.Username = ""
+			_, err = authorizer.Authorize(context.WithValue(ctx, ContextUser, builtinRole))
+			require.NoError(t, err)
+		})
 	}
-
-	// Apply a node lock.
-	nodeLock, err := types.NewLock("node-lock", types.LockSpecV2{
-		Target: types.LockTarget{Node: builtinRole.Identity.Username},
-	})
-	require.NoError(t, err)
-	upsertLockWithPutEvent(ctx, t, srv, nodeLock)
-
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, builtinRole))
-	require.Error(t, err)
-	require.True(t, trace.IsAccessDenied(err))
-
-	builtinRole.Identity.Username = ""
-	_, err = srv.Authorizer.Authorize(context.WithValue(ctx, ContextUser, builtinRole))
-	require.NoError(t, err)
 }
 
-func upsertLockWithPutEvent(ctx context.Context, t *testing.T, srv *TestAuthServer, lock types.Lock) {
-	lockWatch, err := srv.LockWatcher.Subscribe(ctx)
+func upsertLockWithPutEvent(ctx context.Context, t *testing.T, client *testClient, watcher *services.LockWatcher, lock types.Lock) {
+	lockWatch, err := watcher.Subscribe(ctx)
 	require.NoError(t, err)
 	defer lockWatch.Close()
 
-	require.NoError(t, srv.AuthServer.UpsertLock(ctx, lock))
+	require.NoError(t, client.UpsertLock(ctx, lock))
 	select {
 	case event := <-lockWatch.Events():
 		require.Equal(t, types.OpPut, event.Type)
@@ -172,4 +228,85 @@ func upsertLockWithPutEvent(ctx context.Context, t *testing.T, srv *TestAuthServ
 	case <-time.After(2 * time.Second):
 		t.Fatal("Timeout waiting for lock put.")
 	}
+}
+
+type testClient struct {
+	services.ClusterConfiguration
+	services.Trust
+	services.Access
+	services.Identity
+	types.Events
+}
+
+func newTestResources(t *testing.T) (*testClient, *services.LockWatcher, Authorizer) {
+	backend, err := memory.New(memory.Config{})
+	require.NoError(t, err)
+
+	t.Cleanup(func() {
+		require.NoError(t, backend.Close())
+	})
+
+	clusterConfig, err := local.NewClusterConfigurationService(backend)
+	require.NoError(t, err)
+	caSvc := local.NewCAService(backend)
+	accessSvc := local.NewAccessService(backend)
+	identitySvc := local.NewIdentityService(backend)
+	eventsSvc := local.NewEventsService(backend)
+
+	client := &testClient{
+		ClusterConfiguration: clusterConfig,
+		Trust:                caSvc,
+		Access:               accessSvc,
+		Identity:             identitySvc,
+		Events:               eventsSvc,
+	}
+
+	// Set default singletons
+	ctx := context.Background()
+	client.SetAuthPreference(ctx, types.DefaultAuthPreference())
+	client.SetClusterAuditConfig(ctx, types.DefaultClusterAuditConfig())
+	client.SetClusterNetworkingConfig(ctx, types.DefaultClusterNetworkingConfig())
+	client.SetSessionRecordingConfig(ctx, types.DefaultSessionRecordingConfig())
+
+	lockSvc := local.NewAccessService(backend)
+
+	lockWatcher, err := services.NewLockWatcher(ctx, services.LockWatcherConfig{
+		ResourceWatcherConfig: services.ResourceWatcherConfig{
+			Component: "test",
+			Client:    client,
+		},
+		LockGetter: lockSvc,
+	})
+	require.NoError(t, err)
+
+	authorizer, err := NewAuthorizer(clusterName, client, lockWatcher)
+	require.NoError(t, err)
+
+	return client, lockWatcher, authorizer
+}
+
+func createUserAndRole(client *testClient, username string, allowedLogins []string, allowRules []types.Rule) (types.User, types.Role, error) {
+	ctx := context.Background()
+	user, err := types.NewUser(username)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	role := services.RoleForUser(user)
+	role.SetLogins(types.Allow, allowedLogins)
+	if allowRules != nil {
+		role.SetRules(types.Allow, allowRules)
+	}
+
+	err = client.UpsertRole(ctx, role)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	user.AddRole(role.GetName())
+	err = client.UpsertUser(user)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	return user, role, nil
 }
